requirements.txt
aiohttp==3.7.4.post0
aiolimiter==1.0.0b1
asgiref==3.4.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
Django==3.2.7
djangorestframework==3.12.4
idna==3.2
multidict==5.1.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.6.3
